requirements.txt
Flask
flask-cors
numpy
opencv-python-headless
tensorflow
Pillow
gunicorn